fixed multiple option bug.
so i have managed to fix the issue i had with the multiple options being selected which isn't good at all. i think the only things i need now to perfect this my implementation is sending data to the last activity which would be the score and your name as well.