fixed the issue of sending data to the final activity.
so i have managed to fix the issue of solving sending data to the final activity. turned out to be a trivial problem. happy i figured it out on my own. makes me feel good about myself. so what exactly was the problem? so what happened was that i was sending data to the MainActivity.kt from MainActivity0point1.kt meanwhile where that data was needed was in the QuestionActivity.kt and ultimately the FinalActivity.kt. so instead of removing the MainActivity.kt entirely, i just decided to capture the data i was sending there using the getStringExtra method and then send it off to the next activity which is the QuestionActivity using the putExtra method. it worked. i can now say i understand well enough how sending data to different activities work. only lapse would still be sending data from back to front rather than front to back which i know how to do well.
what i wanted to do was to send the data off directly to the final activity rather than going through the two activities in between but i don't know if that is possible yet.